+++ Nextcloud Cookbook iOS Client/AppState.swift
