+++ Nextcloud Cookbook iOS Client/Util/Alerts.swift
