+++ Nextcloud Cookbook iOS Client/Util/DurationComponents.swift
