+++ Nextcloud Cookbook iOS Client/Util/SupportedLanguage.swift
